Removed file route for script tag, removed unnecessary code.
