+++ app/guid-file/route.ts
@@ -9,7 +9,6 @@ import moment from 'moment';
 
 import Institution from 'ember-osf-web/models/institution';
 import MetaTags, { HeadTagDef } from 'ember-osf-web/services/meta-tags';
-import ScriptTags from 'ember-osf-web/services/script-tags';
 import Ready from 'ember-osf-web/services/ready';
 import OsfStorageFile from 'ember-osf-web/packages/files/osf-storage-file';
 import BitbucketFile from 'ember-osf-web/packages/files/bitbucket-file';
@@ -26,19 +25,16 @@ import S3File from 'ember-osf-web/packages/files/s3-file';
 import CurrentUserService from 'ember-osf-web/services/current-user';
 import RegistrationModel from 'ember-osf-web/models/registration';
 import CustomFileMetadataRecordModel from 'ember-osf-web/models/custom-file-metadata-record';
-import CustomItemMetadataRecordModel from 'ember-osf-web/models/custom-item-metadata-record';
 
 export default class GuidFile extends Route {
     @service('head-tags') headTagsService!: HeadTagsService;
     @service intl!: Intl;
     @service metaTags!: MetaTags;
-    @service scriptTags!: ScriptTags;
     @service ready!: Ready;
     @service currentUser!: CurrentUserService;
 
     headTags?: HeadTagDef[];
     metadata!: CustomFileMetadataRecordModel;
-    parentMetadata!: CustomItemMetadataRecordModel;
 
     @task
     @waitFor
@@ -58,84 +54,7 @@ export default class GuidFile extends Route {
                 this.intl.t('general.presented_by_osf'),
             language: this.metadata.language ? this.metadata.language : undefined,
         };
-
-        // Google Structured Data
-        const id = await model.get('id');
-        const parentId = await model.target.get('id');
-        this.parentMetadata = await this.store.findRecord('custom-item-metadata-record', parentId);
-        const file = await this.store.findRecord('file', id, {include: 'target'});
-        const target = await file.target as unknown as RegistrationModel;
-        const contributors = await target.get('contributors');
-        const firstAuthor = contributors[0];
-
-        const funders = await this.parentMetadata.funders;
-        let funderDS = '"funder": [';
-        if (funders) {
-            for (const funder of funders) {
-                const funderName = funder.funder_name;
-                funderDS +=
-                    `
-                        {
-                            "@type": "${this.intl.t('general.structured_data.funder_type_organization')}",
-                            "name": "${funderName}"
-                        }
-                    `;
-                funderDS += ',';
-                funderDS += '\n';
-            }
-            const lastComma = funderDS.lastIndexOf(',');
-            funderDS = funderDS.slice(0, lastComma - 1);
-            funderDS += ']';
-        }
-
-        const fileModelOverrides = `
-            {
-                "context": "${this.intl.t('general.structured_data.context')}",
-                "@type": "${this.intl.t('general.structured_data.dataset')}",
-                "name": "${metaTagsData.title}",
-                "description": "${metaTagsData.description}",
-                "url": "${model.links.self}",
-                "isBasedOn": "${model.links.relationships.target.links.related.href}",
-                "isAccessibleForFree": true,
-                "creator": {
-                    "@type": "${this.intl.t('general.structured_data.creator_type_person')}",
-                    "name": "${firstAuthor}"
-                },
-                ${funderDS},
-                "distribution": [
-                    {
-                        "@type": "${this.intl.t('general.structured_data.distribution_type')}",
-                        "encodingFormat": "${this.intl.t('general.structured_data.encoding_format_md5')}",
-                        "contentUrl": "${model.links.download}"
-                    },
-                    {
-                        "@type": "${this.intl.t('general.structured_data.distribution_type')}",
-                        "encodingFormat": "${this.intl.t('general.structured_data.encoding_format_sha2')}",
-                        "contentUrl": "${model.links.download}"
-                    }
-                ]
-            }
-        `;
-
-        const jsonLD: object | undefined = await this.scriptTags.returnStructuredData(id);
-        const jsonString: string = jsonLD ?
-            JSON.stringify(jsonLD) : fileModelOverrides;
-        const scriptTagData = {
-            type: 'application/ld+json',
-            content: jsonString,
-        };
-
-        const metaTags: HeadTagDef[] = this.metaTags.getHeadTags(metaTagsData);
-        const scriptTag: HeadTagDef[] = this.scriptTags.getHeadTags(scriptTagData);
-        const allTags: HeadTagDef[] = metaTags.concat(scriptTag);
-
-        // Concatenate meta and script tags if both, only meta tags or none
-        if (metaTags && scriptTag) {
-            this.set('headTags', allTags);
-        } else if (metaTags) {
-            this.set('headTags', metaTags);
-        }
-        // Rebuild head tags and clear blocker
+        this.set('headTags', this.metaTags.getHeadTags(metaTagsData));
         this.headTagsService.collectHeadTags();
         blocker.done();
     }
