--- app/services/script-tags.ts
@@ -1,82 +1,0 @@
-import Service, { inject as service } from '@ember/service';
-import config from 'ember-get-config';
-import Intl from 'ember-intl/services/intl';
-import CurrentUserService from 'ember-osf-web/services/current-user';
-import captureException from 'ember-osf-web/utils/capture-exception';
-
-export type Content = object | string | undefined;
-
-export type DataContent = Content | Content[];
-
-export interface ScriptTagsData {
-    type?: DataContent;
-    content?: DataContent;
-}
-
-export interface ScriptTagAttrs {
-    type: Content;
-}
-
-// ember-cli-meta-tags element types
-export enum TagType {
-    SCRIPT = 'script',
-    NOSCRIPT = 'noscript',
-}
-
-export interface HeadTagDef {
-    type: string;
-    content: DataContent;
-    attrs: ScriptTagAttrs;
-}
-
-/**
- *
- * Creates one script tag with model-specific overrides or default attribute values.
- * Should additional script tags be needed, getHeadTags() should be called for each
- * script and added to the headTags array.
- *
- * @method getScriptTagAttributes
- * @method getHeadTags
- * @param {ScriptTagsData} scriptTagsOverrides Overrides defaults
- * @return {HeadTagsDefs} getHeadTags() returns HeadTagDefs for ScriptTagAttrs
- */
-export default class ScriptTags extends Service {
-    @service intl!: Intl;
-    @service currentUser!: CurrentUserService;
-
-    async returnStructuredData(guid: string): Promise<any> {
-        const path = `${config.OSF.url}/${guid}/metadata/?format=google-dataset-json-ld`;
-        let jsonLD;
-        let jsonFetch: object | undefined;
-        try {
-            jsonFetch = await this.currentUser.authenticatedAJAX({
-                method: 'GET',
-                url: path,
-            });
-            if (jsonFetch) {
-                jsonLD = jsonFetch;
-            }
-        } catch (e) {
-            const errorMessage = this.intl.t('general.structured_data.json_ld_retrieval_error');
-            captureException(e, { errorMessage });
-        }
-        return jsonLD;
-    }
-
-    /**
-     * Creates one HTML head element script tag with content and a data MIME type
-     *
-     * @method getHeadTags
-     * @param {ScriptTagsData} scriptTagsData Override values from node
-     * @return {HeadTagDef[]}
-     */
-    getHeadTags(scriptTagsOverrides: ScriptTagsData): HeadTagDef[] {
-        return [{ type: 'script', content: scriptTagsOverrides.content, attrs: { type: scriptTagsOverrides.type } }];
-    }
-}
-
-declare module '@ember/service' {
-    interface Registry {
-        'script-tags': ScriptTags;
-    }
-}
