+++ lib/registries/addon/engine.js
@@ -17,7 +17,6 @@ const engine = Engine.extend({
             'features',
             'head-data',
             'meta-tags',
-            'script-tags',
             'head-tags',
             'intl',
             'media',
